Proyecto Academia Conquerblocks/blog/translation.py
from modeltranslation.translator import translator, TranslationOptions
from .models import Post

class PostTranslationOptions(TranslationOptions):
    fields = ('title', 'content')

translator.register(Post, PostTranslationOptions)