Proyecto Academia Conquerblocks/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
da-vinci==0.4.0
Django==5.0.7
django-ckeditor==6.7.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-js-asset==2.2.0
django-modeltranslation==0.19.9
django-rosetta==0.10.1
django-thumbnails==0.8.0
idna==3.10
pillow==10.4.0
polib==1.2.0
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.1
urllib3==2.2.3
